pyproject.toml
[project]
name = "blender_isometric_scene_setup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = []

[dependency-groups]
dev = [
    "fake-bpy-module-latest>=20260220",
]

[[tool.mypy.overrides]]
module = "bpy"
ignore_missing_imports = true
